+++ .eslintrc.js
@@ -35,10 +35,6 @@ module.exports = {
       rules: {
         '@typescript-eslint/array-type': ['error', {default: 'generic'}],
         '@typescript-eslint/ban-types': 'error',
-        '@typescript-eslint/no-implicit-any-catch': [
-          'error',
-          {allowExplicitAny: true},
-        ],
         '@typescript-eslint/no-unused-vars': [
           'error',
           {argsIgnorePattern: '^_'},