+++ packages/jest-circus/src/types.ts
@@ -7,7 +7,7 @@
 
 // Used as type
 import type {Circus} from '@jest/types';
-import expect = require('expect');
+import type expect = require('expect');
 
 export const STATE_SYM = Symbol(
   'JEST_STATE_SYMBOL',